JSON: Use info.Fields instead of Type.Fields
Previously using Type.Fields would unravel
the fields of embedded structs.

Also fixed up a recursion in decodeReflectJSONInterface
where the concrete type info was only retrieved iff
info.Registered || info.AlwaysDisambiguate but in
cases e.g with an embedded struct of this form
```go
type Rat struct {
    *big.Rat
}
```

despite registration, the condition would never match
sending us through an endless recursion in which
we could never retrieve the concrete type. Since by that
point we are dealing with an interface{}, always disambiguate
and retrieve its concrete type.
